+++ autosub/__init__.py
@@ -7,6 +7,8 @@ Defines autosub's commandline entry point functionality.
 from __future__ import absolute_import, print_function, unicode_literals
 import os
 import gettext
+import sys
+import subprocess
 
 # Import third-party modules
 import pysubs2
@@ -35,7 +37,21 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
     Run autosub as a command-line program.
     """
 
-    args = options.get_cmd_args()
+    is_pause = False
+
+    try:
+        input_main = raw_input
+    except NameError:
+        input_main = input
+
+    option_parser = options.get_cmd_parser()
+    if len(sys.argv) > 1:
+        args = option_parser.parse_args()
+    else:
+        option_parser.print_help()
+        new_argv = input_main(_("\nInput args(without \"autosub\"): "))
+        args = option_parser.parse_args(new_argv.split())
+        is_pause = True
 
     if args.https_proxy:
         os.environ['https_proxy'] = args.https_proxy
@@ -57,10 +73,7 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
             raise exceptions.AutosubException(_("\nAll works done."))
 
         if not args.yes:
-            try:
-                input_m = raw_input
-            except NameError:
-                input_m = input
+            input_m = input_main
         else:
             input_m = None
 
@@ -155,15 +168,15 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
                                     fps=fps,
                                     styles_list=None)
 
+        raise exceptions.AutosubException(_("\nAll works done."))
+
     except KeyboardInterrupt:
         print(_("\nKeyboardInterrupt. Works stopped."))
-        return 1
     except pysubs2.exceptions.Pysubs2Error:
         print(_("\nError: pysubs2.exceptions. Check your file format."))
-        return 1
     except exceptions.AutosubException as err_msg:
         print(err_msg)
-        return 0
 
-    print(_("\nAll works done."))
+    if is_pause:
+        input_main(_("Press Enter to exit..."))
     return 0